refactor(preset): make the legacy-read island a physical boundary
Relocate the island members that lived outside `src/features/preset/` under
live-sounding names into the island, so its boundary is a directory rather
than a comment and a #376 sunset cannot strand them (#387).

- `sequencer/lib/migrations.ts` -> `document/legacy-knob-migrators.ts`
- `transport/lib/legacy-swing.ts` -> `document/legacy-swing.ts`
- `legacyCycleToChain` extracted from `sequencer/lib/chain.ts` into
  `document/legacy-cycle-to-chain.ts` (live `appendChainDraftStep` stays)
- `PresetFileV1` and its island-only satellites (`TransportParams`,
  `SequencerData`, `VariationCycle`) moved into `types/legacy-v1.ts` and
  renamed loud (`LegacyTransportParams`, `LegacySequencerData`,
  `LegacyVariationCycle`); `PresetListItem` and `VARIATION_LABELS` stay live

Stop re-exporting the three version-namespace symbols
(`PRESET_FILE_VERSION`, `isReadablePresetFileVersion`,
`migratePresetFileVersion`) from the live `document/index.ts` barrel; the sole
cross-module reader imports `PRESET_FILE_VERSION` from the module path.

Add `document/versions.ts`, a single readable-version registry
(`READABLE_V1_FILE_VERSIONS`, `PRESET_FILE_VERSION`,
`READABLE_DOCUMENT_VERSION_V2`) that the decode ladder and v1 read helpers
reference in place of scattered literals (#380).

Update docs/preset-versioning.md section 4 to reflect the new physical layout,
the added members, and the dispatch boundaries.

Zero behavior change: all moves/imports/docs, no logic edits.

Closes #387
Closes #380
