+++ docs/preset-versioning.md
@@ -92,13 +92,21 @@ A schema tighten with no bump still requires a fixture proving no in-the-wild fi
 ## 4. The frozen legacy-read island and its sunset seam
 
 The frozen legacy-read island is every surface that reads a pre-canonical (knob or position) value and converts it with a PERMANENTLY pinned curve, plus the tolerant legacy schemas and the one-time storage adopters.
+The boundary is physical, not a comment: every member lives under `src/features/preset/` (in `document/`, `types/legacy-v1.ts`, `session/`, and `library/`), and no live-path module imports any of them except the four dispatch boundaries below (#387).
 Its members:
 
 - `document/migrate-v1.ts` (`frozenV1Curves`: v1 knob -> canonical) and `document/frozen-split-filter.ts` (position -> canonical), the frozen curves.
 - `document/file-v1.ts`, `document/parse.ts`, `document/migrate.ts`, the tolerant v1-family read and normalization.
-- `types/legacy-v1.ts`, the knob-space types kept isolated so the shape cannot leak into live code.
+- `document/legacy-knob-migrators.ts`, the runtime knob-space v1 migrators (pattern / instrument / master-chain), relocated from `features/sequencer/lib/migrations.ts` so its innocuous name can no longer be mistaken for the sequencer's own migrations (#387).
+- `document/legacy-swing.ts`, the pre-#269 swing knob migrator, relocated from `features/transport/lib/legacy-swing.ts` (#387).
+- `document/legacy-cycle-to-chain.ts`, the v1 `variationCycle` -> canonical chain conversion, extracted from `features/sequencer/lib/chain.ts` so it no longer shares a module with the live `appendChainDraftStep` (#387).
+- `document/versions.ts`, the readable-version registry (`READABLE_V1_FILE_VERSIONS`, `PRESET_FILE_VERSION`, `READABLE_DOCUMENT_VERSION_V2`) that the decode ladder and v1 helpers reference in place of scattered literals (#380).
+- `types/legacy-v1.ts`, the knob-space types kept isolated so the shape cannot leak into live code; it holds the flagship `PresetFileV1` and its satellites `LegacyTransportParams`, `LegacySequencerData`, and `LegacyVariationCycle` (relocated from `types/preset.ts`, `features/transport/types/transport.ts`, and `features/sequencer/types/sequencer.ts`, #387).
 - `session/legacy-adopter.ts`, `session/legacy-preset-meta-capture.ts`, and `library/adoption.ts`, the one-time localStorage adopters.
 - `document/__fixtures__/` and `lib/serialization/__fixtures__/` with `corpus.test.ts`, `migrate-v1.golden.ts`, and the parity tests, which pin the frozen curves against the retired live curves.
+
+The four dispatch boundaries where a live-path read enters the island are `document/decode.ts` (the version ladder), `lib/serialization/index.ts` (share links), `session/bootstrap.ts` (adopted-session load), and `library/adoption.ts` (legacy-array adoption); every other island member is imported only by another island member.
+The three version-namespace symbols `PRESET_FILE_VERSION`, `isReadablePresetFileVersion`, and `migratePresetFileVersion` are NO LONGER re-exported from the live `document/index.ts` barrel; the only cross-module reader (`session/legacy-adopter.ts`) imports `PRESET_FILE_VERSION` from `document/versions.ts` directly (#387).
 
 The share compact codec is NO LONGER a member.
 It once held a v1.5 knob-space decoder (`lib/serialization/compact.ts` + `decode.ts`) that fed the same migration ladder; #373 deleted it and made the share codec latest-only, so no share-link ingress reads a knob or position value anymore.
@@ -110,10 +118,11 @@ Stores, the engine, egress, and every newly written file or link are canonical o
 The frozen curves never read a live UI curve; the parity tests assert frozen equals live today, and on a deliberate retune the TEST is updated to pin the frozen values, never the frozen module.
 
 The sunset seam: the island is deletable as one unit at the moment the product accepts breaking a legacy generation.
-The v1.x sub-island (`migrate-v1.ts`, `file-v1.ts`, `parse.ts`, `migrate.ts`, `frozenV1Curves`, `types/legacy-v1.ts`, and the v1 fixtures) can be deleted together once no v1 / v1.5 `.dh` files need to load AND the one-time localStorage adopters have run everywhere and their legacy keys are gone.
+The v1.x sub-island (`migrate-v1.ts`, `file-v1.ts`, `parse.ts`, `migrate.ts`, `legacy-knob-migrators.ts`, `legacy-swing.ts`, `legacy-cycle-to-chain.ts`, `frozenV1Curves`, `types/legacy-v1.ts`, and the v1 fixtures) can be deleted together once no v1 / v1.5 `.dh` files need to load AND the one-time localStorage adopters have run everywhere and their legacy keys are gone.
 The share side of that seam is already cut: #373 removed the `v: 1.5` share decoder ahead of the rest, since share links are ephemeral and safe to break early.
-Deleting the remaining sub-island means removing the `1` / `1.5` rungs from `document/decode.ts`, deleting those modules and fixtures, and letting `decode` reject those versions with `UnsupportedVersionError`.
+Deleting the remaining sub-island means removing the `1` / `1.5` rungs from `document/decode.ts`, dropping `READABLE_V1_FILE_VERSIONS` and `PRESET_FILE_VERSION` from `document/versions.ts`, deleting those modules and fixtures, and letting `decode` reject those versions with `UnsupportedVersionError`.
 `frozen-split-filter.ts` OUTLIVES the v1 sub-island because `migrate-v2.ts` also depends on it; it retires only when both v1 AND v2 documents are no longer read.
+`document/versions.ts` likewise OUTLIVES the v1 sub-island, because its `READABLE_DOCUMENT_VERSION_V2` rung serves the v2 read path.
 Because the island is import-isolated and fully covered by the fixture corpus, each deletion is mechanical and the corpus proves nothing live depended on it.
 There is no scheduled sunset today; until one is chosen, the island is load-bearing forever.
 
@@ -127,9 +136,9 @@ No live correctness bug was found that warranted halting the audit.
   So `v: 2` labeled two incompatible payload shapes, and a pre-#361 `v: 2` link decoded to a `CorruptFieldError` (invalid-link toast + init fallback).
   Severity was low: the change happened inside the unreleased #357 epic across a ~1-day window, share links are ephemeral, and the failure was fail-safe.
   #373 resolved it by collapsing to a single share codec and giving its sole canonical shape a fresh, honest version (`3`): version now maps 1:1 to shape, and any `v: 2` payload (scalar or canonical) is refused uniformly with `UnsupportedVersionError` rather than decoded, so the shape collision can never resurface.
-- Readable-version literals are decentralized (no issue filed).
-  The readable-but-not-writable versions (`1`, `1.5`, `2`) appear as bare literals across `decode.ts` (`PRESET_DOCUMENT_VERSION_V2 = 2`), `migrate.ts`, and `file-v1.ts` rather than a single shared registry.
-  This is not a correctness gap; a shared "readable versions" list would make the decode ladder self-documenting, worth folding into the later sweep.
+- Readable-version literals were decentralized: issue [#380](https://github.com/mxfng/drumhaus/issues/380) - RESOLVED (folded into #387).
+  The readable-but-not-writable versions (`1`, `1.5`, `2`) once appeared as bare literals across `decode.ts` (`PRESET_DOCUMENT_VERSION_V2 = 2`), `migrate.ts`, and `file-v1.ts` rather than a single shared registry.
+  This was not a correctness gap; the shared `document/versions.ts` registry now backs the decode ladder and the v1 read helpers, so the readable set lives in one self-documenting place.
 - The Versioning and Migration-path sections of docs/preset-persistence.md are a past-tense design narrative that predates the shipped outcome (no issue filed).
   They describe an integer `CURRENT_VERSION = 2` and a `MIGRATIONS[]` ladder, whereas the code shipped a fractional `2.1` and an explicit if-ladder in `decode.ts`.
   This doc supersedes them as the living policy; the narrative is left as history.
