+++ src/features/preset/document/index.ts
@@ -28,10 +28,5 @@ export {
 } from "./errors";
 export type { PresetDocumentErrorCode } from "./errors";
 export { presetFileV1Schema } from "./file-v1";
-export {
-  PRESET_FILE_VERSION,
-  isReadablePresetFileVersion,
-  migratePresetFileVersion,
-} from "./migrate";
 export { migrateV1ToDocument } from "./migrate-v1";
 export { parsePresetFileV1, validatePresetFileV1 } from "./parse";
