+++ src/features/preset/types/legacy-v1.ts
@@ -11,6 +11,7 @@
  */
 
 import type { InstrumentRole } from "@/core/audio/engine/instrument/types";
+import type { Pattern, PatternChain } from "@/core/audio/engine/pattern-types";
 import type { SampleData } from "@/features/kit/types/sample";
 import type { InlineMeta, Meta } from "./meta";
 
@@ -58,9 +59,50 @@ interface LegacyKnobMasterChainParams {
   highPass?: number;
 }
 
+/** A v1 file's transport section; `swing` is a 0-100 knob position here. */
+interface LegacyTransportParams {
+  bpm: number;
+  swing: number;
+}
+
+/** A v1 file's pre-chain variation cycle; only ever read, for migration. */
+type LegacyVariationCycle = "A" | "B" | "AB" | "AAAB";
+
+/** A v1 file's sequencer section, including the pre-chain variation cycle. */
+interface LegacySequencerData {
+  pattern: Pattern;
+  /** Legacy variation cycle (pre-chain). Only used for migration. */
+  variationCycle?: LegacyVariationCycle;
+  chain: PatternChain;
+  chainEnabled: boolean;
+}
+
+/**
+ * The knob-space v1 `.dh` preset file (versions 1 and 1.5 share this exact
+ * shape; v1.5 marks the #269 swing retune, which reinterpreted the persisted
+ * swing knob value). This is the island's flagship read-only shape: it is
+ * always deconstructed into canonical stores on load and never written back.
+ * Runtime objects are normalized to version 1.5 (migratePresetFileVersion);
+ * version 1 can still appear in objects read verbatim from storage before
+ * normalization.
+ */
+interface PresetFileV1 {
+  kind: "drumhaus.preset";
+  version: 1 | 1.5;
+  meta: Meta;
+  kit: LegacyKitFile;
+  transport: LegacyTransportParams;
+  sequencer: LegacySequencerData;
+  masterChain: LegacyKnobMasterChainParams;
+}
+
 export type {
   LegacyKitFile,
   LegacyKnobInstrumentData,
   LegacyKnobInstrumentParams,
   LegacyKnobMasterChainParams,
+  LegacySequencerData,
+  LegacyTransportParams,
+  LegacyVariationCycle,
+  PresetFileV1,
 };
