--- src/features/transport/types/transport.ts
@@ -1,6 +1,0 @@
-interface TransportParams {
-  bpm: number;
-  swing: number;
-}
-
-export type { TransportParams };
